requirements.txt
kiteconnect==5.0.1
python-dotenv==1.0.1
pandas==2.3.1
matplotlib==3.9.2
